Arreglados Except de endpoints, añadido idExt y arreglado endpoint get-all para home

+++ app/database.py
@@ -52,8 +52,8 @@ class Castmanu:
     async def isAdmin(self, id):
         result = await self.fetch("SELECT * FROM users WHERE id = %s AND admin = 1", (id))
         if not result:
-            raise HTTPException(status_code=401, detail="No autorizado: no eres administrador")
-        return True
+            return {"success": False, "message": "No tienes permisos de administrador"}
+        return {"success": True}
 
     
     async def register(self, username: str, email: str, password: str, admin: bool = False):
@@ -70,9 +70,44 @@ class Castmanu:
         return {"success": True, "message": "Usuario registrado con éxito."}
     
     # Esto devuelve todas las pelis y demas
-    async def get_all(self):
-        # Primero conseguimos todos los elementos
-        films = await self.fetch("SELECT id, title, type, poster FROM films", (), True)
+    async def get_all(self, pagina, titulo, tipo, generos):
+        offset = (pagina - 1) * 20
+        params = []
+        where_clauses = []
+        join_clause = ""
+        having_clause = ""
+        group_by = ""
+        print(generos)
+        
+        if generos:
+            join_clause = "INNER JOIN film_genres ON IdFilm = id"
+            genre_placeholders = ", ".join(["%s"] * len(generos))
+            where_clauses.append(f"IdGenre IN ({genre_placeholders})")
+            having_clause = f"HAVING COUNT(DISTINCT IdGenre) = {len(generos)}"
+            group_by = "GROUP BY id"
+            params.extend(generos)
+
+        if titulo:
+            where_clauses.append("LOWER(title) LIKE LOWER(%s)")
+            params.append(f"%{titulo}%")
+
+        if tipo:
+            where_clauses.append("type = %s")
+            params.append(tipo)
+
+        where_sql = ""
+        if where_clauses:
+            where_sql = "WHERE " + " AND ".join(where_clauses)
+
+        count_query = f"SELECT CEIL(COUNT(*) / 20.0) AS total FROM (SELECT id FROM films {join_clause} {where_sql} {group_by} {having_clause}) AS sub"
+        total_paginas = await self.fetch(count_query, tuple(params))
+
+        query = f"SELECT id, title, type, poster FROM films {join_clause} {where_sql} {group_by} {having_clause} ORDER BY title LIMIT 20 OFFSET %s"
+        print(query)
+        print(count_query)
+        params.append(offset)
+        films = await self.fetch(query, tuple(params), True)
+
         genres_map = await self.fetch("SELECT idFilm, genre FROM film_genres INNER JOIN genres ON idGenre = id", (), True)
 
         # Agrupamos los géneros por id de película
@@ -84,7 +119,7 @@ class Castmanu:
         for film in films:
             film["genres"] = generos_por_pelicula.get(film["id"], [])
 
-        return films
+        return {"peliculas": films, "total_paginas": total_paginas["total"]}
     
     # Esto devuelve la info de 1 peli concreta
     async def get_film(self,id):
@@ -104,15 +139,29 @@ class Castmanu:
         return film
     
     # Añadir pelicula o elemento
-    async def add_film(self, title, type, sinopsis, poster_format, uploader, capitulo, generos):
+    async def add_film(self, title, type, sinopsis, poster_format, uploader, capitulo, generos, idExt):
         
-        await self.isAdmin(uploader)
+        admin = await self.isAdmin(uploader)
+
+        if not admin["success"]:
+            return admin
 
         if type == "Serie":
+            if (idExt):
+                existe = await self.fetch("SELECT * FROM films WHERE idExt = %s AND capitulo = %s",(idExt, capitulo))
+            else:
+                existe = await self.fetch("SELECT * FROM films WHERE title = %s AND capitulo = %s",(title, capitulo))
             file = f'https://castmanu.ddns.net/videos/{title}/{capitulo}/master.m3u8'
         else:
+            if (idExt):
+                existe = await self.fetch("SELECT * FROM films WHERE idExt = %s",(idExt))
+            else:
+                existe = await self.fetch("SELECT * FROM films WHERE title = %s",(title))
             file = f'https://castmanu.ddns.net/videos/{title}/master.m3u8'
 
+        if existe:
+            return {"success": False, "message": "Lo que intentas añadir ya existe"}
+        
         poster = None
         if poster_format:
             if poster_format.startswith("http"):
@@ -120,20 +169,24 @@ class Castmanu:
             else:
                 poster = f'https://castmanu.ddns.net/fotos/{title}.{poster_format}'
 
-        id = await self.fetch("INSERT INTO films(title, type, sinopsis, poster, file, uploader, capitulo) VALUES(%s, %s, %s, %s, %s, %s, %s) RETURNING id",(title, type, sinopsis, poster, file, uploader, capitulo))
+        id = await self.fetch("INSERT INTO films(title, type, sinopsis, poster, file, uploader, capitulo, idExt) VALUES(%s, %s, %s, %s, %s, %s, %s, %s) RETURNING id",(title, type, sinopsis, poster, file, uploader, capitulo, idExt))
         if generos:
             insertGeneros = ""
             for genero in generos:
                 insertGeneros += f" ({id["id"]}, {self.mapGenero(genero)}),"
             insertGeneros = insertGeneros[:-1]
             await self.fetch(f"INSERT INTO film_genres VALUES{insertGeneros}")
-        return {"success": True, "message": "Pelicula añadida"}
+        return {"success": True, "message": "Pelicula añadida", "id": id["id"]}
     
     # Eliminar pelicula o elemento
     async def delete_film(self, id, deleter):
 
-        await self.isAdmin(deleter)
+        admin = await self.isAdmin(deleter)
 
+        if not admin["success"]:
+            return admin
+        
+        await self.fetch("DELETE FROM film_genres WHERE idFilm = %s",(id))
         await self.fetch("DELETE FROM films WHERE id = %s",(id))
 
         return {"success": True, "message": "Pelicula eliminada"}
